requirements.txt
Flask
gunicorn
langchain
openai==0.27.0
tiktoken
urllib3
bs4
selenium
unstructured
python-decouple
nltk
python-magic
python-dotenv
pypdf